[feat] 라우터를 위한 정적 데이터 분리 - @/data/board.ts로 이동 - @/types/board.ts로 이동
