+++ src/types/board.ts
@@ -1,0 +1,6 @@
+interface CategoryType {
+  category: string;
+  subclass: string;
+}
+
+export type { CategoryType };
